# compute isogeny kernel polynomials from irreducible divisors or (possibly irrational) points